jobs/generate-thumbnail/requirements.txt
functions-framework==3.*
google-cloud-storage
Pillow
